+++ services/ptr.go
@@ -1,3 +1,0 @@
-package services
-
-func stringPtr(v string) *string { return &v }